requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-extensions==2.1.3
django-filter==2.0.0
djangorestframework==3.8.2
drfdocs==0.0.11
graphviz==0.9
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
openapi-codec==1.3.2
pydot==1.2.4
pyparsing==2.2.2
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0